model/requirements.txt
certifi==2026.2.25
charset-normalizer==3.4.6
contourpy==1.3.3
cycler==0.12.1
easyocr==1.7.2
filelock==3.25.2
fonttools==4.62.1
fsspec==2026.2.0
idna==3.11
ImageIO==2.37.3
Jinja2==3.1.6
kiwisolver==1.5.0
lazy-loader==0.5
MarkupSafe==3.0.3
matplotlib==3.10.8
mpmath==1.3.0
mysql-connector-python==9.6.0
networkx==3.6.1
ninja==1.13.0
numpy==2.4.3
opencv-python==4.13.0.92
opencv-python-headless==4.13.0.92
packaging==26.0
pillow==12.1.1
polars==1.39.3
polars-runtime-32==1.39.3
psutil==7.2.2
pyclipper==1.4.0
pyparsing==3.3.2
python-bidi==0.6.7
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
PyYAML==6.0.3
requests==2.33.0
scikit-image==0.26.0
scipy==1.17.1
setuptools==81.0.0
shapely==2.1.2
six==1.17.0
sympy==1.14.0
tifffile==2026.3.3
torch==2.11.0
torchvision==0.26.0
typing_extensions==4.15.0
ultralytics==8.4.27
ultralytics-thop==2.0.18
urllib3==2.6.3
